2023 04 14 external debug

+++ contracts/interpreter/ops/rain/ISaleV2/OpISaleV2SaleStatus.sol
@@ -13,7 +13,7 @@ library OpISaleV2SaleStatus {
     using LibIntegrityCheck for IntegrityCheckState;
 
     function f(uint256 sale_) internal view returns (uint256) {
-        return uint(ISaleV2(address(uint160(sale_))).saleStatus());
+        return uint256(ISaleV2(address(uint160(sale_))).saleStatus());
     }
 
     function integrity(
